Implemented minimal router for updating map view detail pannel. - moved default views and urls to async folders
- refactored everything to play nicely with django-leaflet
