+++ cadasta/core/static/js/smap/map.js
@@ -1,5 +1,6 @@
-var SMap = (function() {
-  var map = L.map('mapid');
+// var map = L.map('mapid');
+
+var SMap = function(map) {
   var layerscontrol = L.control.layers().addTo(map);
 
   var geojsonTileLayer = new L.TileLayer.GeoJSON(
@@ -15,16 +16,16 @@ var SMap = (function() {
           layer.bindPopup("<div class=\"text-wrap\">" +
                       "<h2><span>Location</span>" +
                       feature.properties.type + "</h2></div>" +
-                      "<div class=\"btn-wrap\"><a href='" + feature.properties.url + "' class=\"btn btn-primary btn-sm btn-block\">" + options.trans['open'] + "</a>"  +
+                      "<div class=\"btn-wrap\"><a href='#/" + feature.properties.url + "' id=\"spatial-pop-up\" class=\"btn btn-primary btn-sm btn-block\">" + options.trans.open + "</a>"  +
                       "</div>");
         }
       }
-    });
+  });
 
   function add_tile_layers() {
     for (var i = 0, n = layers.length; i < n; i++) {
-      var attrs = L.Util.extend(layers[i]['attrs']);
-      var layer = {name: layers[i]['label'], url: layers[i]['url'], options: attrs};
+      var attrs = L.Util.extend(layers[i].attrs);
+      var layer = {name: layers[i].label, url: layers[i].url, options: attrs};
       var l = L.tileLayer(layer.url, layer.options);
       layerscontrol.addBaseLayer(l, layer.name);
 
@@ -54,17 +55,20 @@ var SMap = (function() {
       );
       boundary.addTo(map);
       projectBounds = boundary.getBounds();
+      options.projectExtent = projectBounds;
       if (options.fitBounds === 'project') {
         map.fitBounds(projectBounds);
+        return projectBounds;
       }
     } else {
       map.fitBounds([[-45.0, -180.0], [45.0, 180.0]]);
     }
   }
 
-  load_project_extent()
+  load_project_extent();
 
-  function load_features() {;
+  // *** CURRENTLY DOES NOT WORK ***
+  function load_features() {
     if (options.fitBounds === 'locations') {
       var bounds = geojsonTileLayer.geojsonLayer.getBounds();
       if (bounds.isValid()) {
@@ -77,7 +81,7 @@ var SMap = (function() {
 
   function render_spatial_resource(){
     $.ajax(fetch_spatial_resources).done(function(data){
-      if (data.length == 0) return;
+      if (data.length === 0) return;
       var spatialResources = {};
       $.each(data, function(idx, resource){
         var name = resource.name;
@@ -85,15 +89,15 @@ var SMap = (function() {
         var group = new L.LayerGroup();
         $.each(resource.spatial_resources, function(i, spatial_resource){
           var layer = L.geoJson(spatial_resource.geom).addTo(group);
-          layers['name'] = spatial_resource.name;
-          layers['group'] = group;
+          layers.name = spatial_resource.name;
+          layers.group = group;
         });
         spatialResources[name] = layers;
       });
       $.each(spatialResources, function(sr){
         var layer = spatialResources[sr];
-        layerscontrol.addOverlay(layer['group'], layer['name'], sr);
-      })
+        layerscontrol.addOverlay(layer.group, layer.name, sr);
+      });
     });
   }
 
@@ -102,18 +106,43 @@ var SMap = (function() {
   function geoLocate() {
     return function(event) {
       map.locate({ setView: true });
-    }
+    };
   }
 
-  $(window).on('hashchange', function() {
-    if (window.location.hash === '#overview')
-      $('.content-single').removeClass('detail-hidden')
-    else {
-        $('.content-single').addClass('detail-hidden')
-    }
+  function add_map_controls() {
+    var geocoder = L.control.geocoder('search-QctWfva', {
+      markers: false
+    }).addTo(map);
+    geocoder.on('select', function (e) {
+      map.setZoomAround(e.latlng, 9);
+    });
 
-    window.setTimeout(function() {
-      map.invalidateSize();
-    }, 400);
-  })
-})();
+    var Geolocate = L.Control.extend({
+      options: {
+        position: 'topleft'
+      },
+
+      onAdd: function() {
+        var controlDiv = L.DomUtil.create(
+          'div', 'leaflet-bar leaflet-control leaflet-control-geolocate'
+        );
+        controlDiv.title = gettext('Go to my location');
+        L.DomEvent
+         .addListener(controlDiv, 'click', L.DomEvent.stopPropagation)
+         .addListener(controlDiv, 'click', L.DomEvent.preventDefault)
+         .addListener(controlDiv, 'click', geoLocate(map));
+
+        L.DomUtil.create('span', 'glyphicon glyphicon-map-marker', controlDiv);
+
+        return controlDiv;
+      }
+    });
+
+    map.addControl(new Geolocate());
+    return map;
+  }
+
+  return {
+    add_map_controls: add_map_controls,
+  };
+};
