+++ cadasta/organization/serializers.py
@@ -11,6 +11,9 @@ from core.form_mixins import SuperUserCheck
 from core.serializers import DetailSerializer, FieldSelectorSerializer
 from accounts.models import User
 from accounts.serializers import UserSerializer
+from spatial.serializers import SpatialUnitSerializer
+from party.serializers import PartySerializer
+from resources.serializers import ResourceSerializer
 from .models import Organization, Project, OrganizationRole, ProjectRole
 from .forms import create_update_or_delete_project_role
 
@@ -109,6 +112,17 @@ class ProjectSerializer(DetailSerializer, serializers.ModelSerializer):
             organization_id=organization.id,
             **validated_data
         )
+
+
+class ProjectStateSerializer(DetailSerializer, serializers.ModelSerializer):
+    spatial_units = SpatialUnitSerializer(many=True, read_only=True)
+    parties = PartySerializer(many=True, read_only=True)
+    resources = ResourceSerializer(many=True, read_only=True)
+
+    class Meta:
+        model = Project
+        fields = ('id', 'description', 'urls', 'contacts',
+                  'spatial_units', 'parties', 'resources')
 
 
 class ProjectGeometrySerializer(geo_serializers.GeoFeatureModelSerializer):
