+++ cadasta/spatial/forms.py
@@ -61,8 +61,8 @@ class TenureRelationshipForm(AttributeForm):
     party_type = forms.ChoiceField(required=False, choices=[])
     tenure_type = forms.ChoiceField(required=True, choices=[])
 
-    class Media:
-        js = ('/static/js/rel_tenure.js', '/static/js/party_attrs.js')
+    # class Media:
+    #    js = ('/static/js/rel_tenure.js', '/static/js/party_attrs.js')
 
     def __init__(self, project, spatial_unit, *args, **kwargs):
         super().__init__(*args, **kwargs)
