+++ cadasta/spatial/models.py
@@ -102,13 +102,13 @@ class SpatialUnit(ResourceModelMixin, RandomIDModel):
 
     def get_absolute_url(self):
         return iri_to_uri(reverse(
-            'locations:detail',
+            'organization:project-dashboard',
             kwargs={
                 'organization': self.project.organization.slug,
                 'project': self.project.slug,
-                'location': self.id,
+                # 'location': self.id,
             },
-        ))
+        ) + '#/records/location/' + self.id)
 
 
 def reassign_spatial_geometry(instance):
