+++ cadasta/spatial/urls/async.py
@@ -5,23 +5,45 @@ from ..views import async
 
 urls = [
     url(
-        r'^$',
-        async.SpatialUnitList.as_view(),
-        name='list'),
-    url(
-        r'^tiled/(?P<z>[0-9]+)/(?P<x>[0-9]+)/(?P<y>[0-9]+)/$',
+        r'^spatial/tiled/(?P<z>[0-9]+)/(?P<x>[0-9]+)/(?P<y>[0-9]+)/$',
         async.SpatialUnitTiles.as_view(),
         name='tiled'),
+    url(
+        r'^records/location/new/$',
+        async.LocationsAdd.as_view(),
+        name='add'),
+    url(
+        r'^records/location/(?P<location>[-\w]+)/edit/$',
+        async.LocationEdit.as_view(),
+        name='edit'),
+    url(
+        r'^records/location/(?P<location>[-\w]+)/delete/$',
+        async.LocationDelete.as_view(),
+        name='delete'),
+    url(
+        r'^records/location/(?P<location>[-\w]+)/$',
+        async.LocationDetail.as_view(),
+        name='detail'),
+
+    url(
+        r'^records/location/(?P<location>[-\w]+)/resources/add/$',
+        async.LocationResourceAdd.as_view(),
+        name='resource_add'),
+    url(
+        r'^records/location/(?P<location>[-\w]+)/resources/new/$',
+        async.LocationResourceNew.as_view(),
+        name='resource_new'),
+
+    url(
+        r'^records/location/(?P<location>[-\w]+)/relationships/new/$',
+        async.TenureRelationshipAdd.as_view(),
+        name='relationship_add'),
 ]
 
 
 urlpatterns = [
     url(
         r'^organizations/(?P<organization>[-\w]+)/projects/'
-        '(?P<project>[-\w]+)/spatial/',
+        '(?P<project>[-\w]+)/',
         include(urls)),
-    # url(urls.tilepath(r'^organizations/(?P<organization>[-\w]+)/projects/'
-    #     '(?P<project>[-\w]+)/spatial/'),
-    #     async.SpatialUnitList.as_view(),
-    #     name='location-tiles'),
 ]
