+++ cadasta/templates/party/relationship_delete.html
@@ -1,14 +1,10 @@
-{% extends "party/relationship_detail.html" %}
 {% load i18n %}
 
-{% block modals %}
-
 <form method="POST" class="modal-backdrop">
-  <div class="modal show" data-backdrop="true">
     <div class="modal-dialog">
       <div class="modal-content">
         <div class="modal-header">
-          <a href="{% url 'parties:relationship_detail' object.organization.slug object.slug relationship.id %}" class="close" aria-label="Close"><span aria-hidden="true">&times;</span></a>
+          <a href="#/records/relationship/{{ relationship.id }}" class="close" aria-label="Close"><span aria-hidden="true">&times;</span></a>
           <h3 class="modal-title">{% trans "Delete relationship" %}</h3>
         </div>
         <div class="modal-body">
@@ -16,14 +12,11 @@
         </div>
         <div class="modal-footer">
           <button type="submit" value="Confirm" class="btn btn-danger archive-final pull-right">{% trans "Yes, delete this relationship" %}</button>
-          <a href="{% url 'parties:relationship_detail' object.organization.slug object.slug relationship.id %}"  class="btn btn-link cancel">
+          <a href="#/records/relationship/{{ relationship.id }}"  class="btn btn-link cancel">
             {% trans "Cancel" %}
           </a>
           {% csrf_token %}
         </div>
       </div><!-- /.modal-content -->
     </div><!-- /.modal-dialog -->
-  </div><!-- /.modal -->
 </form>
-
-{% endblock %}
