+++ cadasta/templates/party/relationship_edit.html
@@ -1,39 +1,12 @@
-{% extends "spatial/location_wrapper.html" %}
 {% load i18n %}
 {% load widget_tweaks %}
 {% load filters %}
 {% load staticfiles %}
 
-{% block extra_head %}
-{{ block.super }}
-<link rel="stylesheet" href="{% static 'css/jquery-ui.min.css' %}">
-<link rel="stylesheet" href="{% static 'css/jquery-ui.structure.min.css' %}">
-<link rel="stylesheet" href="{% static 'css/jquery-ui.theme.min.css' %}">
-{% endblock %}
-
-{% block extra_script %}
-{{ block.super }}
-<script src="{% static 'js/jquery-ui.min.js' %}"></script>
-{% if get_current_language != "en-us" %}
-{% get_current_language as LANGUAGE_CODE %}
-<script src="https://cdn.rawgit.com/jquery/jquery-ui/1.12.1/ui/i18n/datepicker-{{ LANGUAGE_CODE }}.js"></script>
-{% endif %}
-<script>
-  $(document).ready(function () {
-    $('.datepicker').datepicker({
-      yearRange: "c-200:c+200",
-      changeMonth: true,
-      changeYear: true,
-    });
-  });
-</script>
-{% endblock %}
-
 
 {% block location_panel %}
-
-<div class="col-md-4 detail detail-edit">
-  <h2><a href="{% url 'locations:detail' object.organization.slug object.slug location.id %}"><span>{% trans "Location" %} </span>{{ location.name }}</a></h2>
+<section>
+  <h2><a href="#/records/location/{{location.id}}"><span>{% trans "Location" %} </span>{{ location.name }}</a></h2>
   <hr class="border-btm">
   <h3>{% trans "Edit Relationship" %}</h3>
   <div class="panel panel-default">
@@ -62,10 +35,9 @@
       </div>
       <div class="panel-footer panel-buttons">
         <button type="submit" class="btn btn-primary">{% trans "Save" %}</button>
-        <a href="{% url 'parties:relationship_detail' object.organization.slug object.slug relationship.id %}" class="btn btn-default">{% trans "Cancel" %}</a>
+        <a href="#/records/relationship/{{ relationship.id }}" class="btn btn-default">{% trans "Cancel" %}</a>
       </div>
     </form>
   </div>
-</div>
-
+</section>
 {% endblock %}
