+++ cadasta/templates/resources/modal_add_lib.html
@@ -21,7 +21,7 @@
           </div>
 
           <div class="modal-footer">
-            <button type="submit" class="btn btn-primary pull-right" name="submit">
+            <button type="submit" class="btn btn-primary pull-right" name="submit" formaction="{{ submit_url }}">
               {% trans "Save" %}
             </button>
              <a class="btn btn-link cancel" href="{{ cancel_url }}{{ cancel_url_hash }}">
