+++ cadasta/templates/spatial/location_add.html
@@ -1,14 +1,10 @@
-{% extends "organization/project_wrapper.html" %}
 {% load i18n %}
 {% load widget_tweaks %}
-
-{% block body-class %} map{% endblock %}
-{% block left-nav %}map{% endblock %}
 
 {% load staticfiles %}
 {% load leaflet_tags %}
 
-{% block page_title %}{% trans "Add location" %} | {% endblock %}
+
 
 {% block extra_head %}
 {% leaflet_css plugins="draw,forms,groupedlayercontrol,markercluster,deflate"%}
@@ -19,58 +15,18 @@
 <link rel="stylesheet" href="{% static 'css/jquery-ui.theme.min.css' %}">
 {% endblock %}
 
+{% block body-class %}map{% endblock %}
 {% block extra_script %}
 {{ block.super }}
 {% leaflet_js plugins="draw,forms,groupedlayercontrol,markercluster,deflate" %}
 <script src="https://cdnjs.cloudflare.com/ajax/libs/leaflet-geocoder-mapzen/1.4.0/leaflet-geocoder-mapzen.js"></script>
 <script src="{% static 'js/jquery-ui.min.js' %}"></script>
+
 {% if get_current_language != "en-us" %}
-{% get_current_language as LANGUAGE_CODE %}
-{{ LANGUAGE_CODE }}
-<script src="https://cdn.rawgit.com/jquery/jquery-ui/1.12.1/ui/i18n/datepicker-{{ LANGUAGE_CODE }}.js"></script>
+  {% get_current_language as LANGUAGE_CODE %}
+  {{ LANGUAGE_CODE }}
+  <script src="https://cdn.rawgit.com/jquery/jquery-ui/1.12.1/ui/i18n/datepicker-{{ LANGUAGE_CODE }}.js"></script>
 {% endif %}
-<script src="{% static 'js/map_utils.js' %}"></script>
-<script>
-  $(document).ready(function () {
-    $(window).on('map:init', function(e) {
-      console.log(e);
-      if (e.detail) {
-        var detail = e.detail;
-      } else {
-        var detail = e.originalEvent.detail;
-      }
-      var options = detail.options;
-      add_map_controls(detail.map);
-      switch_layer_controls(detail.map, options);
-
-      var trans = {
-        open: "{% trans 'Open location' %}"
-      };
-
-      {% if object.extent %}
-      var projectExtent = {{ object.extent.geojson|safe }};
-      {% else %}
-      var projectExtent = null;
-      {% endif %}
-
-      renderFeatures(detail.map,
-                     '{% url "async:spatial:list" object.organization.slug object.slug %}',
-                     {projectExtent: projectExtent, trans: trans, fitBounds: 'project'});
-
-      var orgSlug = '{{ object.organization.slug }}';
-      var projectSlug = '{{ object.slug }}';
-      var url = '/api/v1/organizations/'
-              + orgSlug + '/projects/' + projectSlug + '/spatialresources/';
-      add_spatial_resources(detail.map, url);
-    });
-
-    $('.datepicker').datepicker({
-      yearRange: "c-200:c+200",
-      changeMonth: true,
-      changeYear: true,
-    });
-  });
-</script>
 {% endblock %}
 
 {% block content %}
