+++ cadasta/templates/spatial/location_form.html
@@ -2,51 +2,38 @@
 {% load widget_tweaks %}
 {% load filters %}
 
-<form method="POST" id="location-wizard" class="col-sm-12 content-single" data-parsley-validate data-parsley-excluded=":hidden, [disabled]">
-  {% csrf_token %}
-  <div class="row">
-    <!-- Start overview map  -->
-    <div class="col-md-8 map">
-      {{ form.geometry }}
+<h2><span>{% trans "Location" %} </span>{% if add == 'yes' %}{% trans "Add location" %}{% else %}{% trans "Edit details" %}{% endif %}</h2>
+<div class="panel panel-default">
+  <div class="panel-body">
+    <h3>{% trans "Draw location on map" %}</h3>
+    <div class="clearfix form-group{% if form.geometry.errors %} has-error{% endif %}">
+      <label class="control-label required">{% trans "1. Use the tools provided on the left side of the map to mark your new location." %}</label>
+      <div class="error-block">{{ form.geometry.errors }}</div>
     </div>
-    <!-- / end overview map -->
-    <!-- Overview detail  -->
-    <div class="col-md-4 detail detail-edit">
-      <h2><span>{% trans "Location" %} </span>{% if add == 'yes' %}{% trans "Add location" %}{% else %}{% trans "Edit details" %}{% endif %}</h2>
-      <div class="panel panel-default">
-        <div class="panel-body">
-          <h3>{% trans "Draw location on map" %}</h3>
-          <div class="clearfix form-group{% if form.geometry.errors %} has-error{% endif %}">
-            <label class="control-label required">{% trans "Use the tools provided on the left side of the map to mark your new location." %}</label>
-            <div class="error-block">{{ form.geometry.errors }}</div>
-          </div>
-          <h3>{% trans "Overview" %}</h3>
-          <div class="form-group{% if form.type.errors %} has-error{% endif %}">
-            <label class="control-label" for="{{ form.type.id_for_label }}" {{ form.type.field.labels_xlang|safe }}>{% trans "What type of location is this?" %}</label>
-            {% render_field form.type class+="form-control" data-parsley-required="true" %}
-            <div class="error-block">{{ form.type.errors }}</div>
-          </div>
-          {% for field in form %}
-            {% if "spatialunit::" in field.name %}
-            <div class="form-group{% if field.errors %} has-error{% endif %}">
-              {{ field.label_xlang }}
-              <label class="control-label" for="{{ field.id_for_label }}" {{ field.field.labels_xlang|safe }}>{{ field.label }}</label>
-              {% if field|field_type == "datefield" %}
-                {{ field|add_class:"form-control datepicker"|set_parsley_required }}
-              {% else %}
-                {{ field|add_class:"form-control"|set_parsley_required }}
-              {% endif %}
-              <div class="error-block">{{ field.errors }}</div>
-            </div>
-            {% endif %}
-          {% endfor %}
-        </div>
-        <div class="panel-footer panel-buttons">
-          <input class="btn btn-primary" type="submit" value="{% trans 'Save' %}"/>
-          <a class="btn btn-default cancel" href="{{ cancel_url }}">{% trans "Cancel" %}</a>
-        </div>
+
+    <h3>{% trans "Overview" %}</h3>
+    <div class="form-group{% if form.type.errors %} has-error{% endif %}">
+      <label class="control-label" for="{{ form.type.id_for_label }}" {{ form.type.field.labels_xlang|safe }}>{% trans "What type of location is this?" %}</label>
+      {% render_field form.type class+="form-control" data-parsley-required="true" %}
+      <div class="error-block">{{ form.type.errors }}</div>
+    </div>
+    {% for field in form %}
+      {% if "spatialunit::" in field.name %}
+      <div class="form-group{% if field.errors %} has-error{% endif %}">
+        {{ field.label_xlang }}
+        <label class="control-label" for="{{ field.id_for_label }}" {{ field.field.labels_xlang|safe }}>{{ field.label }}</label>
+        {% if field|field_type == "datefield" %}
+          {{ field|add_class:"form-control datepicker"|set_parsley_required }}
+        {% else %}
+          {{ field|add_class:"form-control"|set_parsley_required }}
+        {% endif %}
+        <div class="error-block">{{ field.errors }}</div>
       </div>
-    </div>
-    <!-- / overview detail -->
+      {% endif %}
+    {% endfor %}
   </div>
-</form>
+  <div class="panel-footer panel-buttons">
+    <input class="btn btn-primary" type="submit" value="{% trans 'Save' %}"/>
+    <a class="btn btn-default cancel" href="{{ cancel_url }}">{% trans "Cancel" %}</a>
+  </div>
+</div>
