+++ cadasta/templates/spatial/location_wrapper.html
@@ -15,7 +15,7 @@
 <<<<<<< HEAD
 {{ block.super }}
 {% leaflet_js plugins="groupedlayercontrol,markercluster,deflate" %}
-<script src="{% static 'js/map_utils.js' %}"></script>
+<!-- <script src="{% static 'js/map_utils.js' %}"></script> -->
 <!-- Detail extra script-->
 <script>
   function locations_map_init(map, options) {
