+++ cadasta/templates/spatial/relationship_add.html
@@ -1,4 +1,3 @@
-{% extends "spatial/location_detail.html" %}
 {% load i18n %}
 {% load widget_tweaks %}
 {% load filters %}
@@ -6,59 +5,16 @@
 
 {% block page_title %}{% trans "Add relationship" %} | {% endblock %}
 
-{% block extra_head %}
-{{ block.super }}
-<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/select2/4.0.3/css/select2.min.css" />
-<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/select2-bootstrap-theme/0.1.0-beta.9/select2-bootstrap.min.css" />
-<link rel="stylesheet" href="{% static 'css/jquery-ui.min.css' %}">
-<link rel="stylesheet" href="{% static 'css/jquery-ui.structure.min.css' %}">
-<link rel="stylesheet" href="{% static 'css/jquery-ui.theme.min.css' %}">
-{% endblock %}
 
-{% block location_extra_script %}
-  {{ form.media }}
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/select2/4.0.3/js/select2.min.js"></script>
-  <script src="{% static 'js/jquery-ui.min.js' %}"></script>
-  {% if get_current_language != "en-us" %}
-  {% get_current_language as LANGUAGE_CODE %}
-  <script src="https://cdn.rawgit.com/jquery/jquery-ui/1.12.1/ui/i18n/datepicker-{{ LANGUAGE_CODE }}.js"></script>
-  {% endif %}
-  <script>
-$(document).ready(function() {
-  var template = function(party) {
-    if (!party.id) {
-      return party.text;
-    }
-    return $(
-      '<div class="party-option">' +
-      '<strong class="party-name">' + party.text + '</strong>' +
-      '<span class="party-type">' + party.element.dataset.type + '</span>' +
-      '</div>'
-    );
-  };
-  $("#party-select").select2({
-    minimumResultsForSearch: 6,
-    templateResult: template,
-    theme: "bootstrap",
-  });
-
-  $('.datepicker').datepicker({
-    yearRange: "c-200:c+200",
-    changeMonth: true,
-    changeYear: true,
-  });
-});
-  </script>
-{% endblock %}
 
 {% block form_modal %}
 <div class="modal-backdrop">
   <div class="modal show" data-backdrop="true">
     <div class="modal-dialog">
       <div class="modal-content">
-        <form method="POST" data-parsley-validate data-parsley-excluded=":hidden, disabled">
+        <form method="POST" data-parsley-validate data-parsley-excluded=":hidden, disabled" id="modal-form">
           <div class="modal-header">
-            <a class="close" href="{% url 'locations:detail' object.organization.slug object.slug location.id %}">
+            <a class="close" href="#/records/location/{{location.id}}">
               <span aria-hidden="true">&times;</span>
             </a>
             <h3 class="modal-title">{% trans "Add relationship" %}</h3>
@@ -127,7 +83,7 @@ $(document).ready(function() {
               {% trans "Save" %}
             </button>
             <a class="btn btn-link cancel"
-              href="{% url 'locations:detail' object.organization.slug object.slug location.id %}">
+              href="#/records/location/{{location.id}}">
               {% trans "Cancel" %}
             </a>
           </div>
