+++ cadasta/templates/spatial/resources_add.html
@@ -1,16 +1,11 @@
-{% extends "spatial/location_detail.html" %}
 {% load i18n %}
-{% load staticfiles %}
 
 {% block page_title %}{% trans "Add resource for location" %} | {% endblock %}
 
 {% block location_extra_script %}
-<script src="{% static 'js/dataTables.forms.js' %}"></script>
 {% include 'resources/script_add_lib.html' %}
 {% endblock %}
 
 {% block modals %}
-{% url 'locations:detail' object.organization.slug object.slug location.id as cancel_url %}
-{% url 'locations:resource_new' object.organization.slug object.slug location.id as upload_url %}
-{% include 'resources/modal_add_lib.html' with cancel_url_hash="#resources" %}
+{% include 'resources/modal_add_lib.html' %}
 {% endblock %}
